refactor(frontend): Rename ETH WalletConnect modal
