+++ src/frontend/src/eth/components/wallet-connect/EthWalletConnectSignModal.svelte
